[Solidity] fix codacy issues: remove unused variables and fix build break
Remove unused variables flagged by Codacy static analysis:
- cname in solidity_convert_ref.cpp
- c_type in solidity_convert_type.cpp
- lt_sol in solidity_convert_tuple.cpp
- sol_type_str in solidity_convert_expr.cpp
- name in solidity_convert_constructor.cpp
- debug_modulename in solidity_convert_builtin.cpp

Also fix build break caused by static_lifetime_init being
inaccessible from solidity_convert_contract.cpp after the
frontend file split — moved to solidity_convert.h.

Test results: 339 total, 331 passed, 0 failed, 8 timeout (all THOROUGH)
